# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_httpresp.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  backend_internal_httpresp.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/backend_internal_httpresp.ErrorBody'
    type: object
...
  internal_auth.ChangePasswordRequest:
    properties:
      current_password:
        example: oldpassword123
        maxLength: 72
        type: string
      new_password:
        example: newpassword456
        maxLength: 72
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
  internal_auth.LoginRequest:
    properties:
      password:
        example: securepassword123
        maxLength: 72
        type: string
    required:
    - password
    type: object
  internal_auth.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_at:
        example: 1705276800
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_auth.LogoutResponse:
    properties:
      message:
        example: logged out
        type: string
    type: object
  internal_auth.OnboardingConfigRequest:
    properties:
      anthropic_key:
        example: sk-ant-...
        type: string
      auto_threshold:
        example: 95
        type: integer
      custom_job_sites:
        example:
        - '["https://careers.example.com"]'
        items:
          type: string
        type: array
      job_sources:
        example:
        - '["greenhouse:openai"'
        - '"lever:figma"]'
        items:
          type: string
        type: array
      livekit_key:
        example: APIxxxxxxxxxxxxxxx
        type: string
      livekit_secret:
        example: secretxxxxxxxxxxxxxxxxxxxxxxxxxx
        type: string
      livekit_url:
        example: wss://livekit.example.com
        type: string
      ms_client_id:
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        type: string
      ms_client_secret:
        example: secretxxxxxxxxxxxxxxxxxxxxxxxxxx
        type: string
      ms_tenant_id:
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        type: string
      openai_key:
        example: sk-...
        type: string
      review_threshold:
        example: 80
        type: integer
    type: object
  internal_auth.OnboardingConfigResponse:
    properties:
      message:
        example: configuration saved
        type: string
    type: object
  internal_auth.OnboardingStepRequest:
    properties:
      step:
        enum:
        - llm
        - voice
        - preferences
        - complete
        example: llm
        type: string
    required:
    - step
    type: object
  internal_auth.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        minLength: 64
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.RefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_at:
        example: 1705276800
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_auth.SetupRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: securepassword123
        maxLength: 72
        minLength: 8
        type: string
      username:
        example: john_doe
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_auth.SetupResponse:
    properties:
      message:
        example: setup completed successfully
        type: string
    type: object
  internal_auth.SetupStatusResponse:
    properties:
      setup_required:
        example: true
        type: boolean
      step:
        example: llm
        type: string
    type: object
  internal_auth.TestEmailRequest:
    properties:
      client_id:
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        type: string
      client_secret:
        example: secretxxxxxxxxxxxxxxxxxxxxxxxxxx
        type: string
      tenant_id:
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        type: string
    required:
    - client_id
    - client_secret
    - tenant_id
    type: object
  internal_auth.TestLLMRequest:
    properties:
      api_key:
        example: sk-...
        type: string
      provider:
        enum:
        - openai
        - anthropic
        example: openai
        type: string
    required:
    - api_key
    - provider
    type: object
  internal_auth.TestLLMResponse:
    properties:
      valid:
        example: true
        type: boolean
    type: object
  internal_auth.TestVoiceRequest:
    properties:
      api_key:
        example: APIxxxxxxxxxxxxxxx
        type: string
      api_secret:
        example: secretxxxxxxxxxxxxxxxxxxxxxxxxxx
        type: string
      url:
        example: wss://livekit.example.com
        type: string
    required:
    - api_key
    - api_secret
    - url
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/ai-job-search-agent/issues
  description: Local-first AI-powered job search automation platform. Discovers, scores,
    and applies to jobs with tailored resumes and cover letters.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/ai-job-search-agent
  title: AI Job Search Agent API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the current user's password (requires authentication)
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/internal_auth.OnboardingConfigResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Current password incorrect or unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with password and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login with tokens
          schema:
            $ref: '#/definitions/internal_auth.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke all refresh tokens and increment session version (requires
        authentication)
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            $ref: '#/definitions/internal_auth.LogoutResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New access and refresh tokens
          schema:
            $ref: '#/definitions/internal_auth.RefreshResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /auth/setup:
    post:
      consumes:
      - application/json
      description: Create the first user account and complete initial setup
      parameters:
      - description: Setup credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.SetupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Setup completed successfully
          schema:
            $ref: '#/definitions/internal_auth.SetupResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "409":
          description: Setup already completed
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Complete initial setup
      tags:
      - Auth
  /auth/setup/complete-onboarding:
    post:
      consumes:
      - application/json
      description: Mark onboarding as complete and enable full API access
      produces:
      - application/json
      responses:
        "200":
          description: Onboarding completed
          schema:
            $ref: '#/definitions/internal_auth.OnboardingConfigResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete onboarding
      tags:
      - Auth
  /auth/setup/config:
    post:
      consumes:
      - application/json
      description: Save LLM keys, voice config, email config, and preferences during
        onboarding
      parameters:
      - description: Onboarding configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.OnboardingConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration saved
          schema:
            $ref: '#/definitions/internal_auth.OnboardingConfigResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Save onboarding configuration
      tags:
      - Auth
  /auth/setup/onboarding-step:
    post:
      consumes:
      - application/json
      description: Update the current onboarding step to track progress
      parameters:
      - description: Step to set
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.OnboardingStepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Step updated
          schema:
            $ref: '#/definitions/internal_auth.OnboardingConfigResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Update onboarding step
      tags:
      - Auth
  /auth/setup/status:
    get:
      consumes:
      - application/json
      description: Check if initial setup is required and get current onboarding step
      produces:
      - application/json
      responses:
        "200":
          description: Setup status
          schema:
            $ref: '#/definitions/internal_auth.SetupStatusResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Get setup status
      tags:
      - Auth
  /auth/setup/test-email:
    post:
      consumes:
      - application/json
      description: Validate Microsoft Graph email configuration during onboarding
      parameters:
      - description: Email configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.TestEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            $ref: '#/definitions/internal_auth.TestLLMResponse'
        "400":
          description: Invalid request body or tenant ID
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Test email configuration
      tags:
      - Auth
  /auth/setup/test-llm:
    post:
      consumes:
      - application/json
      description: Validate an LLM provider API key during onboarding
      parameters:
      - description: LLM provider and API key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.TestLLMRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            $ref: '#/definitions/internal_auth.TestLLMResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Test LLM API key
      tags:
      - Auth
  /auth/setup/test-voice:
    post:
      consumes:
      - application/json
      description: Validate LiveKit voice configuration during onboarding
      parameters:
      - description: Voice configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.TestVoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            $ref: '#/definitions/internal_auth.TestLLMResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Test voice configuration
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...